Migrate test_interest_over_time to vcr
Now we can test the returned value by the backend, however we don't
compare the _full_ response because it's a `pd.DataFrame` with 52 rows,
instead we just compare the length, the first 3 rows and the last 3
rows, and from that we can infer that the result is correct.

Later we can add more fine-grained tests if we want but I think this is
a good starting point.