Migrate request_args passing tests to respones
Here we too a different approach because there is no observable
behavior in the result, in this case is in the requests made.
We switch to the responses library to test the exact requests made.

Also we do it with `trending_searches` because it's the easiest of all
calls to mock, and because all the code passed through `_get_data` which
is the one who make the requests.

Also we wrote just one test instead of two, there's no point in testing
only two methods, we can either test only one and hope for the best or
test all.
While testing all whould be better is also a time consuming chore with
very little value.